Fix some warnings in mod_secure_download.c
git-svn-id: svn://svn.lighttpd.net/lighttpd/trunk@2566 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/mod_secure_download.c
@@ -275,10 +275,10 @@ URIHANDLER_FUNC(mod_secdownload_uri_handler) {
 	}
 
 	/* timed-out */
-	if ( (srv->cur_ts > ts && srv->cur_ts - ts > p->conf.timeout) ||
-	     (srv->cur_ts < ts && ts - srv->cur_ts > p->conf.timeout) ) {
+	if ( (srv->cur_ts > ts && srv->cur_ts - ts > (int) p->conf.timeout) ||
+	     (srv->cur_ts < ts && ts - srv->cur_ts > (int) p->conf.timeout) ) {
 		if (con->conf.log_request_handling) {
-			TRACE("timestamp is too old: %ld, timeout: %d", ts, p->conf.timeout);
+			TRACE("timestamp is too old: %ld, timeout: %d", (long int) ts, p->conf.timeout);
 		}
 
 		/* "Gone" as the url will never be valid again instead of "408 - Timeout" where the request may be repeated */
